Update Mod-Sketch.ino
Added Toy.setID(1); else it wont work ..
Added command for get Powerlevel - which was missing, it was only possible to decrease/increase and get the current Power this way
Added command to return Version info, hard to detect a Dilduino until now
Reworked the big if { } else if { } else if { } ... into a switch command. (Some switch-cases require additonal { }, not sure why, but works this way)
Added some more documentation.